feat: 실 좌석 API 연동 및 더미 데이터 제거 refactor: 좌석 선택·주문·결제 플로우를 seatId 기반 구조로 전환

+++ spring-chat-matching/src/main/java/com/chatmatchingservice/springchatmatching/domain/event/dto/SectionResponseDto.java
@@ -7,7 +7,7 @@ public record SectionResponseDto(
         String code,           // 115, FLOOR_A
         String name,           // 115구역
         String grade,          // VIP / R / S
-        int price,             // ticket.price
+        Long price,             // ticket.price
         int totalSeats,
         Long remainSeats,
         List<SeatResponseDto> seats
